KNN modificado

+++ src/Lectura.h
@@ -7,6 +7,6 @@ void generar_training(std::vector<Dato>&, const string&);
 void generar_test(std::vector<Dato>&, const string&);
 
 template <typename Iterator>
-void leer_matriz(Mat<uint>& resultado, Iterator& it);
+void leer_matriz(Mat<double>& resultado, Iterator& it);
 
 #endif //METODOS_TP2_LECTURA_H
